Refactor styled Components
