+++ src/components/CategoryFilter.js
@@ -1,18 +1,7 @@
 import React from 'react';
-import styled from 'styled-components';
 import PropTypes from 'prop-types';
+import { Select } from './styled/lib';
 
-const Select = styled.select`
-  width: 180px;
-  height: 45px;
-  padding: 6px 15px 9px 16px;
-  border-radius: 4px;
-  border: solid 1px var(--gray-border);
-  background-color: var(--white-two);
-  &:focus {
-    outline: none;
-  }
-`;
 const categoryList = [
   'Action', 'Biography', 'History', 'Horror', 'Kids', 'Learning', 'Sci-Fi',
 ];
